lightbox: Move api/downloadFile.js to lightbox/downloadImage.js.
Resolve a naming collision in LightboxActionSheet.js by renaming
the function previously named `downloadFile` to `tryToDownloadImage`.
