+++ src/lightbox/downloadImage.js
@@ -2,7 +2,7 @@
 import { CameraRoll, Platform } from 'react-native';
 import RNFetchBlob from 'rn-fetch-blob';
 
-import type { Auth } from './apiTypes';
+import type { Auth } from '../api/apiTypes';
 import { getAuthHeader, getFullUrl } from '../utils/url';
 import userAgent from '../utils/userAgent';
