+++ src/lightbox/shareImage.android.js
@@ -1,5 +1,5 @@
 /* @flow */
-import download from '../api/downloadFile';
+import download from './downloadImage';
 import type { Auth } from '../types';
 import ShareImageAndroid from '../nativeModules/ShareImageAndroid';
