+++ src/lightbox/shareImage.ios.js
@@ -2,12 +2,12 @@
 import { Share } from 'react-native';
 
 import type { Auth } from '../types';
-import downloadFile from '../api/downloadFile';
+import downloadImage from './downloadImage';
 import { showToast } from '../utils/info';
 
 export default async (url: string, auth: Auth) => {
   try {
-    const uri = await downloadFile(url, auth);
+    const uri = await downloadImage(url, auth);
     try {
       await Share.share({ url: uri, message: url });
     } catch (error) {
